+++ NEWS.md
@@ -10,7 +10,8 @@
   * `nv_eye()` to create an identity matrix.
   * `nv_solve()` to solve a system of linear equations.
   * `nv_cholesky()` to compute the Cholesky decomposition of a matrix.
-
+* Improved semantics:
+  * `nvl_cholesky()` now zeros out the upper/lower triangle of the output.
 
 ## Bug Fixes